fixed some Props not working because of missing xCount
